split NotifyJob into NotifyCommentJob and NotifyMessageJob
