+++ spec/controllers/comments_controller_spec.rb
@@ -18,7 +18,7 @@ describe CommentsController do
       stub_login_as reader
       post :create, params: {story_id: story.short_id, comment: "great story!"}
       expect(response.status).to eq(302)
-      expect(NotifyJob).to have_been_enqueued.exactly(:once)
+      expect(NotifyCommentJob).to have_been_enqueued.exactly(:once)
     end
   end
 end
